+++ Scripts/Installers/UITemplateAdsInstaller.cs
@@ -14,7 +14,6 @@ namespace TheOneStudio.UITemplate.UITemplate.Installers
 
         public override void InstallBindings()
         {
-            this.Container.BindInterfacesAndSelfTo<UITemplateAutoHideMRECService>().AsCached().NonLazy();
             this.Container.BindInterfacesAndSelfTo<UITemplateAnalyticHandler>().AsCached();
 #if CREATIVE
             this.Container.Bind<UITemplateAdServiceWrapper>().To<UITemplateAdServiceWrapperCreative>().AsCached();
