+++ Scripts/ThirdPartyServices/UITemplateAdServiceWrapperCreative.cs
@@ -4,6 +4,7 @@
     using System.Collections.Generic;
     using Core.AdsServices;
     using GameFoundation.Scripts.UIModule.ScreenFlow.BaseScreen.Presenter;
+    using GameFoundation.Scripts.UIModule.ScreenFlow.Managers;
     using GameFoundation.Scripts.Utilities.LogService;
     using ServiceImplementation.Configs;
     using ServiceImplementation.Configs.Ads;
@@ -15,10 +16,9 @@
     {
         public UITemplateAdServiceWrapperCreative(ILogService logService, AdServicesConfig adServicesConfig, SignalBus signalBus, IAdServices adServices, List<IMRECAdService> mrecAdServices,
             UITemplateAdsController uiTemplateAdsController, UITemplateGameSessionDataController gameSessionDataController, List<IAOAAdService> aoaAdServices, IBackFillAdsService backFillAdsService,
-            ToastController toastController, UITemplateLevelDataController levelDataController, ThirdPartiesConfig thirdPartiesConfig,
-            UITemplateAutoHideMRECService autoHideMrecService)
+            ToastController toastController, UITemplateLevelDataController levelDataController, ThirdPartiesConfig thirdPartiesConfig, IScreenManager screenManager)
             : base(logService, adServicesConfig, signalBus, adServices, mrecAdServices, uiTemplateAdsController,
-                gameSessionDataController, aoaAdServices, backFillAdsService, toastController, levelDataController, thirdPartiesConfig, autoHideMrecService)
+                gameSessionDataController, aoaAdServices, backFillAdsService, toastController, levelDataController, thirdPartiesConfig, screenManager)
         {
         }
